refactor(ingestion/task): extract index-doc mapping into task/indexdoc package
Move the pipeline-output -> search-engine index document mapping helpers out
of the task package into a dedicated, dependency-light leaf package
internal/ingestion/task/indexdoc. These functions (NormalizeChunks,
ProcessChunksForPipeline, AddPositions, AggregateTableDocMetadata,
GetEmbeddingTokenConsumption, DeepCopyChunks and the table-column config
resolvers) only depend on common/utility and are not task-orchestration
concerns, so they now form a single cohesive unit that can be reused and
tested without pulling in the task package's dao/engine/service graph.

- chunk_utils.go -> indexdoc/normalize.go (deepCopyChunks exported as
  DeepCopyChunks; used by golden_compare.go)
- chunk_process.go -> indexdoc/process.go
- position.go -> indexdoc/position.go
- EmbeddingTokenConsumptionKey constant moved into indexdoc/constants.go
- task/constants.go keeps only GRAPH_RAPTOR_FAKE_DOC_ID
- Call sites in pipeline_executor.go and golden_compare.go now reference
  the indexdoc package; package-task tests qualify the moved symbols.

This is a pure mechanical relocation: no behavior change. Unit tests for
both packages remain green.
